backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.41
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.0.0
openai==1.12.0